requirements.txt
Django==3.2.18
python-telegram-bot==13.3
environs==9.5.0
python-dateutil==2.8.2
gunicorn==20.0.4
django_debug_toolbar==3.8.1
django-admin-extra-buttons==1.5.6
